+++ src/api/workspace/__tests__/del.test.js
@@ -1,4 +1,4 @@
-import del from '../delete'
+import del from '../del'
 
 jest.mock('../../call', () => ({
   call: async () => 'res'
